Skip download file if already exists in /doc directory

+++ main.php
@@ -28,10 +28,11 @@ foreach ($reports as $research) {
     $row['authors'] = Authors::repack(trim($research->{'@Author'}));
     $row['published_date'] = substr($research->{'@CompletionDate'}, 0, 10);
     $row['content'] = null;
-    $row['doc_url'] = trim($research->FilePath);
+    $row['doc_url'] = (isset($research->FilePath)) ? trim($research->FilePath) : null;
     $rows[$research_no] = $row;
 }
 
+$TIKA_URL = Config::get('TIKA_URL');
 foreach ($rows as $row) {
     // skip row without doc_url
     $doc_url = $row['doc_url'];
@@ -44,13 +45,16 @@ foreach ($rows as $row) {
     $dot_idx = strrpos($doc_url, '.');
     $file_extension = mb_substr($doc_url, $dot_idx);
     $save_file_at = $doc_base . $research_no . $file_extension;
-    file_put_contents($save_file_at, file_get_contents($doc_url));
+    if (! file_exists($save_file_at)) {
+        file_put_contents($save_file_at, file_get_contents($doc_url));
+    }
 
     // convert file html via tika
     $save_html_at = $html_base . $research_no . '.html';
-    $TIKA_URL = Config::get('TIKA_URL');
-    $command = sprintf("curl --silent -T %s %s -H 'Accept: text/html' --output %s", $save_file_at, $TIKA_URL, $save_html_at);
-    exec($command);
+    if (! file_exists($save_html_at)) {
+        $command = sprintf("curl --silent -T %s %s -H 'Accept: text/html' --output %s", $save_file_at, $TIKA_URL, $save_html_at);
+        exec($command);
+    }
 }
 
 // packing extracted document content into csv
